Move config related functions to new file
Move functions related to loading and validating kube config to separate
file.
